Deprecate screenSize modifiers and environment key while keeping them working #patch
