+++ Sources/OversizeUI/Controls/HUD/HUD.swift
@@ -6,8 +6,6 @@
 import SwiftUI
 
 public struct HUD<Title: View, Icon: View>: View {
-    @Environment(\.screenSize) var screenSize
-
     private let content: HUDContent<Title, Icon>
     private let isAutoHide: Bool
